-Dodano tryb Grupy (2 mecze) + Puchar w trybie offline

+++ cup/views.py
@@ -130,9 +130,10 @@ def dashboard(request, cup_id: int):
     """
     cup = Cup.objects.get(id=cup_id)
     matches_in_cup = Match.objects.filter(cup=cup)
-    if cup.type == 'GrupyPuchar1mecz':
+    match_in_cup_not_played = False
+    if cup.type == 'GrupyPuchar1mecz' or cup.type == 'GrupyPuchar2mecze':
         groups = Cup.objects.filter(groupanothercup=cup)
-        rounds_in_groups = Round.objects.filter(cup__groupanothercup=cup)
+        rounds_in_groups = Round.objects.filter(cup__groupanothercup=cup).order_by('match', 'id')
         matches_in_groups = Match.objects.filter(cup__groupanothercup=cup)
         matches_in_cup = Match.objects.filter(cup=cup)
         match_in_groups_not_played = False
@@ -141,7 +142,6 @@ def dashboard(request, cup_id: int):
                 pass
             else:
                 match_in_groups_not_played = True
-        match_in_cup_not_played = False
         for match in list(matches_in_cup):
             if match.finished:
                 pass
@@ -941,6 +941,231 @@ def generate_round(request, cup_id: int):
                 cup.save()
                 messages.success(request, f'Wygenerowano {cup.actual_round}.')
                 return HttpResponseRedirect('/cup/dashboard/{}/'.format(cup.id))
+    elif cup.type == 'GrupyPuchar2mecze':
+        if not cup.league_generated:
+            players = list(Player.objects.filter(cup=cup))
+            if len(players) > 128:
+                len_group = 32
+            elif len(players) > 64:
+                len_group = 16
+            elif len(players) > 32:
+                len_group = 8
+            elif len(players) > 16:
+                len_group = 4
+            elif len(players) > 8:
+                len_group = 2
+            else:
+                len_group = 2
+            ordering = list(cup.players_order[:-1].split(','))
+            if 'pausing' in ordering:
+                ordering.remove('pausing')
+            print(ordering)
+            print('lengroup' + str(len_group))
+            ordering_in_groups = [[] for _ in range(len_group)]
+            print(ordering_in_groups)
+            len_while = 1
+            all_rounds = 0
+            while ordering:
+                print(ordering)
+                if len_while <= len_group:
+                    ordering_in_groups[len_while - 1].append(ordering.pop(0))
+                    len_while += 1
+                    print(ordering_in_groups)
+                else:
+                    len_while = 1
+            for index, group in enumerate(ordering_in_groups):
+                if len(group) % 2 != 0:
+                    ordering_in_groups[index].append('pausing')
+            for nr_group in range(len_group + 1)[1:]:
+                print(nr_group)
+                if len(ordering_in_groups[nr_group - 1]) % 2 == 0:
+                    all_rounds = len(ordering_in_groups[nr_group - 1])
+                else:
+                    all_rounds = len(ordering_in_groups[nr_group - 1]) + 1
+                print("all_rounds " + str(all_rounds))
+                players = []
+                print(ordering_in_groups)
+                ordering_to_group = ''
+                for pl_id in ordering_in_groups[nr_group - 1]:
+                    ordering_to_group += pl_id + ','
+                ordering_to_group = list(ordering_to_group[:-1].split(','))
+                new_group = Cup.objects.create(type='2 mecze', name=str(str(nr_group) + '_grupa_' + cup.name),
+                                               author=request.user,
+                                               choosing_teams=cup.choosing_teams, online=cup.online,
+                                               declarations=cup.declarations, groupanothercup=cup,
+                                               players_order=ordering_to_group, league_generated=True,
+                                               registration=cup.registration)
+                for player_id in ordering_in_groups[nr_group - 1]:
+                    if player_id:
+                        if player_id != 'pausing':
+                            player_to_add = Player.objects.get(id=player_id)
+                            player_to_add.group = new_group
+                            player_to_add.save()
+                            players.append(player_to_add)
+                        else:
+                            players.append('pausing')
+                    else:
+                        continue
+                for round_nr in range(all_rounds)[1:]:
+                    actual_round = []
+                    if round_nr == 1:
+                        new_round = Round.objects.create(cup=new_group, name=f'Kolejka {round_nr}')
+                        new_round2 = Round.objects.create(cup=new_group, name=f'Kolejka {round_nr + all_rounds - 1}', match=2)
+                        for match in range(len(players) // 2):
+                            if players[0] == 'pausing':
+                                pausing = players.pop(-1)
+                                new_round.pausing = pausing
+                                new_round.save()
+                                players.remove(players[0])
+                                players.remove(pausing)
+                            elif players[-1] == 'pausing':
+                                pausing = players[0]
+                                new_round.pausing = pausing
+                                new_round.save()
+                                players.remove(pausing)
+                                players.remove(players[-1])
+                            else:
+                                pl1 = players.pop(0)
+                                pl2 = players.pop(-1)
+                                actual_round.append([pl1, pl2])
+                        for match in actual_round:
+                            if round_nr % 2 == 0:
+                                Match.objects.create(player1=match[0],
+                                                     player2=match[1],
+                                                     cup=new_group,
+                                                     round=new_round)
+                                Match.objects.create(player1=match[0],
+                                                     player2=match[1],
+                                                     cup=new_group,
+                                                     round=new_round2)
+                            else:
+                                Match.objects.create(player1=match[1],
+                                                     player2=match[0],
+                                                     cup=new_group,
+                                                     round=new_round)
+                                Match.objects.create(player1=match[1],
+                                                     player2=match[0],
+                                                     cup=new_group,
+                                                     round=new_round2)
+                        new_round.league_generated = True
+                        new_round.save()
+                    else:
+                        players = []
+                        new_ordering = [ordering_to_group[0], ordering_to_group[-1]]
+                        for player in ordering_to_group[1:-1]:
+                            new_ordering.append(player)
+                        ordering_to_group = new_ordering
+                        for player_id in ordering_to_group:
+                            if player_id != 'pausing':
+                                player_to_add = Player.objects.get(id=player_id)
+                                player_to_add.group = new_group
+                                player_to_add.save()
+                                players.append(player_to_add)
+                            else:
+                                players.append('pausing')
+                        new_round = Round.objects.create(cup=new_group, name=f'Kolejka {round_nr}')
+                        new_round2 = Round.objects.create(cup=new_group, name=f'Kolejka {round_nr + all_rounds - 1}', match=2)
+                        for match in range(len(players) // 2):
+                            if players[0] == 'pausing':
+                                pausing = players.pop(-1)
+                                players.remove(players[0])
+                                new_round.pausing = pausing
+                                new_round.save()
+                            elif players[-1] == 'pausing':
+                                pausing = players.pop(0)
+                                players.remove(players[-1])
+                                new_round.pausing = pausing
+                                new_round.save()
+                            else:
+                                actual_round.append([players.pop(0), players.pop(-1)])
+                        for match in actual_round:
+                            if round_nr % 2 == 0:
+                                Match.objects.create(player1=match[0],
+                                                     player2=match[1],
+                                                     cup=new_group,
+                                                     round=new_round)
+                                Match.objects.create(player1=match[0],
+                                                     player2=match[1],
+                                                     cup=new_group,
+                                                     round=new_round2)
+                            else:
+                                Match.objects.create(player1=match[1],
+                                                     player2=match[0],
+                                                     cup=new_group,
+                                                     round=new_round)
+                                Match.objects.create(player1=match[1],
+                                                     player2=match[0],
+                                                     cup=new_group,
+                                                     round=new_round2)
+                        new_round.league_generated = True
+                        new_round.save()
+            cup.len_rounds = all_rounds
+            cup.league_generated = True
+            cup.save()
+            messages.success(request, 'Faza grupowa wygenerowana.')
+            return HttpResponseRedirect('/cup/dashboard/{}/'.format(cup.id))
+        else:
+            if not cup.cupgenerated:
+                groups = Cup.objects.filter(groupanothercup=cup)
+                cup = Cup.objects.get(id=cup_id)
+                for group in groups:
+                    players = Player.objects.filter(group=group).order_by('-points')[:2]
+                    for player in players:
+                        group.promotion.add(player)
+                        group.save()
+                        cup.promotion.add(player)
+                        cup.save()
+                cup.cupgenerated = True
+            players = list(cup.promotion.all())
+            print(players)
+            shuffle(players)
+            len_players = 0
+            cup.save()
+            new_round = Round.objects.create(cup=cup)
+            match_not_played = False
+            for match in list(matches):
+                if match.finished:
+                    pass
+                else:
+                    match_not_played = True
+            if match_not_played:
+                messages.error(request, 'Nie rozegrano wszystkich meczy. Wprowadź wszystkie wyniki.')
+                return HttpResponseRedirect(f'/cup/dashboard/{cup.id}/')
+            else:
+                if cup.actual_round is not None:
+                    finished_round = Round.objects.get(id=cup.actual_round.id)
+                    for player_prom in list(finished_round.promotion.all()):
+                        new_round.players.add(player_prom)
+                if cup.actual_round:
+                    players = list(cup.actual_round.promotion.all())
+                if len(players) == 2:
+                    new_round.name = 'Finał'
+                elif len(players) == 4:
+                    new_round.name = 'Półfinał'
+                elif len(players) == 8:
+                    new_round.name = 'Ćwierćfinał'
+                elif len(players) == 16:
+                    new_round.name = '1/8 Finału'
+                elif len(players) == 32:
+                    new_round.name = '1/16 Finału'
+                elif len(players) == 64:
+                    new_round.name = '1/32 Finału'
+                elif len(players) == 128:
+                    new_round.name = '1/64 Finału'
+                for new_match in range(len(players) // 2):
+                    player1 = random.choice(players)
+                    players.remove(player1)
+                    player2 = random.choice(players)
+                    players.remove(player2)
+                    match = Match.objects.create(player1=player1, player2=player2, cup=cup, round=new_round)
+                    new_round.players.add(match.player1)
+                    new_round.players.add(match.player2)
+                new_round.save()
+
+                cup.actual_round = new_round
+                cup.save()
+                messages.success(request, f'Wygenerowano {cup.actual_round}.')
+                return HttpResponseRedirect('/cup/dashboard/{}/'.format(cup.id))
         # if (len(players) // 4) % 2 == 0:
         #     len_group = len(players) // 4
         # else:
@@ -1108,7 +1333,7 @@ def delete_the_result(request, cup_id: int, match_id: int):
     cup = Cup.objects.get(id=cup_id)
     actual_round = cup.actual_round
     match = Match.objects.get(id=match_id)
-    if cup.type == 'Puchar':
+    if cup.type == 'Puchar' or cup.cupgenerated:
         if match.result1 > match.result2:
             actual_round.promotion.remove(match.player1)
             actual_round.save()
@@ -1196,16 +1421,20 @@ def join_the_cup(request, cup_id):
 @login_required
 def left_the_cup(request, cup_id):
     cup = Cup.objects.get(id=cup_id)
-    if request.user in cup.players.all():
-        cup.players.remove(request.user)
-        cup.number_of_players -= 1
-        cup.save()
-        player_to_del = Player.objects.get(user=request.user)
-        player_to_del.delete()
-        messages.success(request, f"Zrezygnowałeś z udziału w {cup.name}.")
-        return HttpResponseRedirect('/cup/online/cup_list_with_open_registration/')
+    if cup.registration == "Otwarta":
+        if request.user in cup.players.all():
+            cup.players.remove(request.user)
+            cup.number_of_players -= 1
+            cup.save()
+            player_to_del = Player.objects.get(user=request.user)
+            player_to_del.delete()
+            messages.success(request, f"Zrezygnowałeś z udziału w {cup.name}.")
+            return HttpResponseRedirect('/cup/online/cup_list_with_open_registration/')
+        else:
+            messages.error(request, f"Nie ma Ciebie na liście rozgrywek {cup.name}, więc nie możesz zrezygnować.")
+            return HttpResponseRedirect('/cup/online/cup_list_with_open_registration/')
     else:
-        messages.error(request, f"Nie ma Ciebie na liście rozgrywek {cup.name}, więc nie możesz zrezygnować.")
+        messages.error(request, "Nie można zrezygnować z rozgrywek po zamknięciu rejestracji.")
         return HttpResponseRedirect('/cup/online/cup_list_with_open_registration/')
